+++ Bazzar/cart-service/src/main/java/ru/bazzar/cart/api/CartItemDto.java
@@ -1,33 +1,31 @@
-package ru.bazzar.api;
+package ru.bazzar.cart.api;
 
 import java.math.BigDecimal;
 
-public class OrderItemDto {
-    private Long id;
+public class CartItemDto {
+    private Long productId;
     private String productTitle;
-    private Long orderId;
     private int quantity;
     private BigDecimal pricePerProduct;
     private BigDecimal price;
 
-    public OrderItemDto() {
+    public CartItemDto() {
     }
 
-    public OrderItemDto(Long id, String productTitle, Long orderId, int quantity, BigDecimal pricePerProduct, BigDecimal price) {
-        this.id = id;
+    public CartItemDto(Long productId, String productTitle, int quantity, BigDecimal pricePerProduct, BigDecimal price) {
+        this.productId = productId;
         this.productTitle = productTitle;
-        this.orderId = orderId;
         this.quantity = quantity;
         this.pricePerProduct = pricePerProduct;
         this.price = price;
     }
 
-    public Long getId() {
-        return id;
+    public Long getProductId() {
+        return productId;
     }
 
-    public void setId(Long id) {
-        this.id = id;
+    public void setProductId(Long productId) {
+        this.productId = productId;
     }
 
     public String getProductTitle() {
@@ -36,14 +34,6 @@ public class OrderItemDto {
 
     public void setProductTitle(String productTitle) {
         this.productTitle = productTitle;
-    }
-
-    public Long getOrderId() {
-        return orderId;
-    }
-
-    public void setOrderId(Long orderId) {
-        this.orderId = orderId;
     }
 
     public int getQuantity() {
@@ -72,13 +62,12 @@ public class OrderItemDto {
 
     @Override
     public String toString() {
-        return "OrderItemDto{" +
-                "id=" + id +
+        return "CartItemDto{" +
+                "productId=" + productId +
                 ", productTitle='" + productTitle + '\'' +
-                ", orderId=" + orderId +
                 ", quantity=" + quantity +
-                ", costPerProduct=" + pricePerProduct +
-                ", cost=" + price +
+                ", pricePerProduct=" + pricePerProduct +
+                ", price=" + price +
                 '}';
     }
 }
