+++ Bazzar/cart-service/src/main/java/ru/bazzar/cart/api/ProductDto.java
@@ -1,4 +1,4 @@
-package ru.bazzar.api;
+package ru.bazzar.cart.api;
 
 import java.math.BigDecimal;
