+++ Bazzar/core-service/src/main/java/ru/bazzar/core/api/AccessException.java
@@ -1,4 +1,4 @@
-package ru.bazzar.api;
+package ru.bazzar.core.api;
 
 public class AccessException extends RuntimeException {
     public AccessException(String message) {
